Move unit test to feature test as it requires a db connection
