+++ tests/Unit/PlanetServiceTest.php
@@ -222,23 +222,4 @@ class PlanetServiceTest extends UnitTestCase
         // Should only return valid buildings, ( ie metal_mine crystal_mine, solar_plant )
         $this->assertEquals(150, $this->planetService->getBuildingCount());
     }
-
-    /**
-     * Tests object building queue status.
-     */
-    public function testIsBuildingObject(): void
-    {
-        $this->createAndSetPlanetModel([
-            'id' => 1,
-        ]);
-
-        // Add level 3 shipyard to building queue
-        $queue = new BuildingQueue();
-        $queue->planet_id = $this->planetService->getPlanetId();
-        $queue->object_id = 21;
-        $queue->object_level_target = 3;
-        $queue->save();
-
-        $this->assertTrue($this->planetService->isBuildingObject('shipyard', 3));
-    }
 }
